Move extractor to the compression module
