+++ pkg/compression/extractor.go
@@ -1,20 +1,18 @@
-package extractor
+package compression
 
 import (
 	"path/filepath"
 
-	"github.com/giongto35/cloud-game/v2/pkg/extractor/zip"
+	"github.com/giongto35/cloud-game/v2/pkg/compression/zip"
 )
 
 type Extractor interface {
 	Extract(src string, dest string) ([]string, error)
 }
 
-const (
-	zipExt = ".zip"
-)
+const zipExt = ".zip"
 
-func NewFromExt(path string) Extractor {
+func NewExtractorFromExt(path string) Extractor {
 	switch filepath.Ext(path) {
 	case zipExt:
 		return zip.New()
