+++ pkg/downloader/pipe/pipe.go
@@ -3,13 +3,13 @@ package pipe
 import (
 	"os"
 
-	"github.com/giongto35/cloud-game/v2/pkg/extractor"
+	"github.com/giongto35/cloud-game/v2/pkg/compression"
 )
 
 func Unpack(dest string, files []string) []string {
 	var res []string
 	for _, file := range files {
-		if unpack := extractor.NewFromExt(file); unpack != nil {
+		if unpack := compression.NewExtractorFromExt(file); unpack != nil {
 			if _, err := unpack.Extract(file, dest); err == nil {
 				res = append(res, file)
 			}
